cleanup | more meaningful filenames, move index.ts into api folder
