+++ api/config/app.config.ts
@@ -2,9 +2,9 @@ import bodyParser from 'body-parser';
 import cors from 'cors';
 import express from 'express';
 import routes from '../routes';
-import corsOptions from '../config/cors.config';
+import corsOptions from './cors.config';
 import errorHandler from '../middlewares/errorHandler';
-import swaggerDocs from './swagger';
+import swaggerDocs from './swagger.config';
 import createHttpError from 'http-errors';
 import 'express-async-errors';
