+++ api/middlewares/__test__/errorHandler.test.ts
@@ -9,7 +9,7 @@ describe('Error Handler', () => {
   it('should include stack in the response in development mode', async () => {
     process.env.NODE_ENV = 'development';
 
-    const { default: app } = require('../../utils/express');
+    const { default: app } = require('../../config/app.config');
 
     const path = '/nope';
     const { statusCode, body } = await supertest(app).get(path);
@@ -21,7 +21,7 @@ describe('Error Handler', () => {
 
   it('should not include stack in the response in production mode', async () => {
     process.env.NODE_ENV = 'production';
-    const { default: app } = require('../../utils/express');
+    const { default: app } = require('../../config/app.config');
 
     const path = '/nope';
     const { statusCode, body } = await supertest(app).get(path);
@@ -33,7 +33,7 @@ describe('Error Handler', () => {
 
   it('should not include stack in the response in test mode', async () => {
     process.env.NODE_ENV = 'test';
-    const { default: app } = require('../../utils/express');
+    const { default: app } = require('../../config/app.config');
 
     const path = '/nope';
     const { statusCode, body } = await supertest(app).get(path);
