+++ thunder-tests/environments/tc_env_pizza-api.json
@@ -12,31 +12,31 @@
         },
         {
             "name": "pizzaId",
-            "value": "64ca8bf989a8eaba87ce95e7"
+            "value": "64cac4908e31e39f5a077f16"
         },
         {
             "name": "pizzaId1",
-            "value": "64ca8b8d89a8eaba87ce95c6"
+            "value": "64cac4908e31e39f5a077f07"
         },
         {
             "name": "pizzaId2",
-            "value": "64ca8b8d89a8eaba87ce95c8"
+            "value": "64cac4908e31e39f5a077f09"
         },
         {
             "name": "pizzaId3",
-            "value": "64ca8b8d89a8eaba87ce95ca"
+            "value": "64cac4908e31e39f5a077f0b"
         },
         {
             "name": "orderId",
-            "value": "64ca8bef89a8eaba87ce95df"
+            "value": "64cac4908e31e39f5a077f1d"
         },
         {
             "name": "orderId1",
-            "value": "64ca8b8d89a8eaba87ce95ce"
+            "value": "64cac4908e31e39f5a077f0f"
         },
         {
             "name": "orderId2",
-            "value": "64ca8b8e89a8eaba87ce95d3"
+            "value": "64cac4908e31e39f5a077f14"
         }
     ]
 }
